Documenting enhancements in read_exercise():
Convenience extension for `cloze` exercises with elements embedded by
`##ANSWERi##` tags: The "Answerlist" only has to be specified if there
are `schoice`/`mchoice` elements, otherwise it can be omitted. For
exercises with mixed elements, the "Answerlist" only has to specify
the entries of the `schoice`/`mchoice` elements while the empty entries
for all other `num` or `string`/`essay`/`file` elements can optionally
be omitted.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/exams@2456 edb9625f-4e0d-4859-8d74-9fd3b1da38cb